requirements.txt
alembic==1.14.0
aniso8601==9.0.1
async-timeout==4.0.3
blinker==1.8.2
cachelib==0.13.0
chardet==5.2.0
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
mysql-connector-python==9.0.0
mysqlclient==2.2.4
ordered-set==4.1.0
packaging==24.1
pillow==10.4.0
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
reportlab==4.2.2
rich==13.8.0
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
wrapt==1.16.0
pyparsing==2.4.7 # sometime it needs to download this package globally out side the virtual environment